+++ src/ToledoMessage.Shared/DTOs/AuthResponse.cs
@@ -1,3 +1,3 @@
 namespace ToledoMessage.Shared.DTOs;
 
-public sealed record AuthResponse(decimal UserId, string Username, string DisplayName, string Token, string? RefreshToken = null);
+public sealed record AuthResponse(long UserId, string Username, string DisplayName, string Token, string? RefreshToken = null);
