+++ src/ToledoMessage.Shared/DTOs/ConversationListItemResponse.cs
@@ -3,7 +3,7 @@ using ToledoMessage.Shared.Enums;
 namespace ToledoMessage.Shared.DTOs;
 
 public sealed record ConversationListItemResponse(
-    decimal ConversationId,
+    long ConversationId,
     ConversationType Type,
     string DisplayName,
     DateTimeOffset? LastMessageTime,
