+++ src/ToledoMessage.Shared/DTOs/CreateConversationRequest.cs
@@ -1,3 +1,3 @@
 namespace ToledoMessage.Shared.DTOs;
 
-public sealed record CreateConversationRequest(decimal ParticipantUserId);
+public sealed record CreateConversationRequest(long ParticipantUserId);
