+++ src/ToledoMessage.Shared/DTOs/UserSearchResult.cs
@@ -2,4 +2,4 @@ namespace ToledoMessage.Shared.DTOs;
 
 public sealed record UserSearchResponse(List<UserSearchResult> Users);
 
-public sealed record UserSearchResult(decimal UserId, string Username, string DisplayName, int DeviceCount);
+public sealed record UserSearchResult(long UserId, string Username, string DisplayName, int DeviceCount);
