+++ src/ToledoMessage/Models/RefreshToken.cs
@@ -4,10 +4,10 @@ public class RefreshToken
 {
     // ReSharper disable  NullableWarningSuppressionIsUsed
 
-    public decimal Id { get; set; }
-    public decimal UserId { get; set; }
+    public long Id { get; set; }
+    public long UserId { get; set; }
     public string Token { get; set; } = string.Empty;
-    public decimal? DeviceId { get; set; }
+    public long? DeviceId { get; set; }
     public DateTimeOffset ExpiresAt { get; set; }
     public DateTimeOffset CreatedAt { get; set; }
     public bool IsRevoked { get; set; }
