+++ src/ToledoMessage/Models/UserPreferences.cs
@@ -3,8 +3,8 @@ namespace ToledoMessage.Models;
 public class UserPreferences
 {
     // ReSharper disable  NullableWarningSuppressionIsUsed
-    public decimal Id { get; set; }
-    public decimal UserId { get; set; }
+    public long Id { get; set; }
+    public long UserId { get; set; }
     public string Theme { get; set; } = "default";
     public string FontSize { get; set; } = "medium";
     public string Language { get; set; } = "en";
